requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
stripe>=7.0.0
httpx>=0.25.0
pyyaml>=6.0.0
prometheus-client>=0.19.0
# Phase 3-4 GTM dependencies
reportlab>=4.0.0  # PDF invoice generation
redis>=5.0.0  # Distributed rate limiting, locking, caching
tiktoken>=0.5.0  # LLM token counting
APScheduler==3.10.4
